Refacturing header and navigation components and add functionality to close pop up affter choise montly year trips
